@@ -1,12 +1,12 @@
 <Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <TargetFramework>net6.0</TargetFramework>
+    <TargetFramework>net8.0</TargetFramework>
     <ImplicitUsings>enable</ImplicitUsings>
     <Nullable>enable</Nullable>
     <GeneratePackageOnBuild>True</GeneratePackageOnBuild>
     <Title>Random sentence</Title>
-    <Version>1.1.1.2</Version>
+    <Version>1.1.1.3</Version>
     <Authors>Kalandor01</Authors>
     <Copyright>MIT</Copyright>
     <PackageProjectUrl>https://github.com/Kalandor01/RandomSentence_CSharp</PackageProjectUrl>
